add template

+++ client/src/view/web/posts/Post.jsx
@@ -53,13 +53,19 @@ const Post = ({ item, type, setRender = () => {}, setShow = () => {} }) => {
 
   const onDelete = async () => {
     if (!userInfo?._id) return onWarning();
-    const response = await deletePostApi({ _id: item._id });
-    if (response) {
-      const response = await getInfoApi();
-      if (response) {
-        setUserInfo(response);
-      } else localStorage.removeItem('token');
-    }
+    showConfirm({
+      title: 'Bạn có chắc chắn muốn xóa bài viết này không!',
+      action: async () => {
+        const response = await deletePostApi({ _id: item._id });
+        if (response) {
+          setRender((pre) => !pre);
+          const response = await getInfoApi();
+          if (response) {
+            setUserInfo(response);
+          } else localStorage.removeItem('token');
+        }
+      }
+    });
   };
 
   return (
